+++ src/solver_advanced/solver/log.rs
@@ -1,9 +1,11 @@
 use std::fs::File;
+use std::fs::OpenOptions;
 use std::io::Write;
 
 use crate::solver_advanced::residuals::ResidualsValues;
 
 extern crate chrono;
+extern crate rustc_version_runtime;
 extern crate whoami;
 use chrono::prelude::*;
 
@@ -14,14 +16,16 @@ const FLOAT_WIDTH: usize = 30;
 const INT_WIDTH: usize = 6;
 
 pub struct SolverLog {
-    content: String,
+    path: String,
 }
 
 /// Log for debugging information
 ///
 /// This object defines the format and concatenate the debugging informations
 impl SolverLog {
-    pub fn new() -> Self {
+    pub fn new(path: &str) -> Self {
+        let mut file = File::create(path).unwrap();
+
         let mut content = String::new();
         content.push_str(&"Runner informations\n");
         content.push_str(&"===================\n\n");
@@ -33,6 +37,9 @@ impl SolverLog {
         content.push_str(&"\n");
         content.push_str(&"Username: ");
         content.push_str(&whoami::username());
+        content.push_str(&"\n");
+        content.push_str("Rust version: ");
+        content.push_str(&rustc_version_runtime::version().to_string());
         content.push_str(&"\n");
         const VERSION: &str = env!("CARGO_PKG_VERSION");
         content.push_str("newton_rootfinder version: ");
@@ -51,15 +58,20 @@ impl SolverLog {
         content.push_str(&"\n");
         content.push_str(&"\n");
 
-        SolverLog { content }
+        write!(file, "{}", content).unwrap();
+
+        SolverLog {
+            path: path.to_string(),
+        }
     }
 
-    pub fn add_content(&mut self, new_content: &str) {
-        self.content.push_str(new_content);
+    pub fn add_content(&self, new_content: &str) {
+        let mut file = OpenOptions::new().append(true).open(&self.path).unwrap();
+        write!(file, "{}", new_content).unwrap();
     }
 
     pub fn add_parameters(
-        &mut self,
+        &self,
         solver_parameters: &str,
         iteratives_config: &str,
         residuals_config: &str,
@@ -70,7 +82,7 @@ impl SolverLog {
     }
 
     pub fn add_damping(
-        &mut self,
+        &self,
         iteratives: &nalgebra::DVector<f64>,
         residuals: &ResidualsValues,
         errors: &nalgebra::DVector<f64>,
@@ -81,7 +93,7 @@ impl SolverLog {
         self.add_iteration(iteratives, residuals, errors);
     }
     pub fn add_new_iteration(
-        &mut self,
+        &self,
         iteratives: &nalgebra::DVector<f64>,
         residuals: &ResidualsValues,
         errors: &nalgebra::DVector<f64>,
@@ -95,7 +107,7 @@ impl SolverLog {
     }
 
     fn add_iteration(
-        &mut self,
+        &self,
         iteratives: &nalgebra::DVector<f64>,
         residuals: &ResidualsValues,
         errors: &nalgebra::DVector<f64>,
@@ -129,10 +141,5 @@ impl SolverLog {
             self.add_content(&entry);
         }
         self.add_content(&"\n");
-    }
-
-    pub fn write(&self, path: &str) {
-        let mut f = File::create(path).unwrap();
-        write!(f, "{}", self.content).unwrap();
     }
 }