+++ tests/log/broyden.rs
@@ -23,13 +23,11 @@ fn broyden_case10_fd() {
         &iteratives,
         &residuals_config,
     );
-    rf.set_debug(true);
+    rf.activate_debug(&LOG_PATH);
 
     let mut user_model = nrf::model::UserModelWithFunc::new(problem_size, broyden1965_case10);
 
     rf.solve(&mut user_model);
-
-    rf.write_log(&LOG_PATH);
 
     let log_ref = File::open(&"./tests/log/log_ref.txt").unwrap();
     let log_new = File::open(&LOG_PATH).unwrap();
@@ -40,7 +38,7 @@ fn broyden_case10_fd() {
     let mut lines_new = log_new_reader.lines();
     let mut lines_ref = log_ref_reader.lines();
 
-    // Parse the runner informations except the time
+    // Parse the runner informations
     for _i in 0..3 {
         let line_new = lines_new.next().unwrap();
         let line_ref = lines_ref.next().unwrap();
@@ -54,6 +52,9 @@ fn broyden_case10_fd() {
     lines_new.next();
     lines_ref.next();
     // ignore the username line
+    lines_new.next();
+    lines_ref.next();
+    // ignore the rustc version line
     lines_new.next();
     lines_ref.next();
     // ignore the crate version line