+++ tests/solver/broyden1_inv_jac/broyden.rs
@@ -16,7 +16,7 @@ fn broyden_case5_fd() {
         init_broyden1965_case5(),
         solution_broyden1965_case5(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -33,7 +33,7 @@ fn broyden_case5_jac() {
         init_broyden1965_case5(),
         solution_broyden1965_case5(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -42,14 +42,14 @@ fn broyden_case5_jac() {
 #[test]
 fn broyden_case6_fd() {
     let problem_size = 5;
-    let damping = true;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case6,
         init_broyden1965_case6(),
         solution_broyden1965_case6(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -58,7 +58,7 @@ fn broyden_case6_fd() {
 #[test]
 fn broyden_case6_jac() {
     let problem_size = 5;
-    let damping = true;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case6,
@@ -66,7 +66,7 @@ fn broyden_case6_jac() {
         init_broyden1965_case6(),
         solution_broyden1965_case6(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -75,14 +75,14 @@ fn broyden_case6_jac() {
 #[test]
 fn broyden_case7_fd() {
     let problem_size = 10;
-    let damping = true;
+    let damping = false;
     run_test_case_fd(
         problem_size,
         broyden1965_case7,
         init_broyden1965_case7(),
         solution_broyden1965_case7(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -91,7 +91,7 @@ fn broyden_case7_fd() {
 #[test]
 fn broyden_case7_jac() {
     let problem_size = 10;
-    let damping = true;
+    let damping = false;
     run_test_case_jac(
         problem_size,
         broyden1965_case7,
@@ -99,7 +99,7 @@ fn broyden_case7_jac() {
         init_broyden1965_case7(),
         solution_broyden1965_case7(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -115,7 +115,7 @@ fn broyden_case8_fd() {
         init_broyden1965_case8(),
         solution_broyden1965_case8(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -132,7 +132,7 @@ fn broyden_case8_jac() {
         init_broyden1965_case8(),
         solution_broyden1965_case8(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -148,7 +148,7 @@ fn broyden_case9_fd() {
         init_broyden1965_case9(),
         solution_broyden1965_case9(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -165,7 +165,7 @@ fn broyden_case9_jac() {
         init_broyden1965_case9(),
         solution_broyden1965_case9(),
         ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+            UpdateQuasiNewtonMethod::BroydenFirstMethod,
         )),
         damping,
     );
@@ -180,7 +180,7 @@ fn broyden_case9_jac() {
 //        broyden1965_case10,
 //        init_broyden1965_case10(),
 //        solution_broyden1965_case10(),
-//        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(UpdateQuasiNewtonMethod::BroydenSecondMethod)),
+//        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(UpdateQuasiNewtonMethod::BroydenFirstMethod)),
 //    );
 //}
 
@@ -193,6 +193,6 @@ fn broyden_case9_jac() {
 //        broyden1965_case10_jac,
 //        init_broyden1965_case10(),
 //        solution_broyden1965_case10(),
-//        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(UpdateQuasiNewtonMethod::BroydenSecondMethod)),
+//        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(UpdateQuasiNewtonMethod::BroydenFirstMethod)),
 //    );
 //}