+++ tests/solver/greenstadt1_jac/polynom.rs
@@ -3,6 +3,7 @@ extern crate newton_rootfinder;
 use newton_rootfinder::solver_advanced as nrf;
 
 use crate::common::{run_test_case_fd, run_test_case_jac};
+
 use nrf::solver::{QuasiNewtonMethod, ResolutionMethod, UpdateQuasiNewtonMethod};
 
 use nrf::test_cases::polynom;
@@ -16,8 +17,8 @@ fn square() {
         polynom::square2,
         nalgebra::DVector::from_vec(vec![1.0]),
         nalgebra::DVector::from_vec(vec![2_f64.sqrt()]),
-        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(
+            UpdateQuasiNewtonMethod::GreenstadtFirstMethod,
         )),
         damping,
     );
@@ -32,8 +33,8 @@ fn root_with_high_derivative() {
         polynom::root_with_high_derivative,
         nalgebra::DVector::from_vec(vec![0.15]),
         nalgebra::DVector::from_vec(vec![0.1]),
-        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(
+            UpdateQuasiNewtonMethod::GreenstadtFirstMethod,
         )),
         damping,
     );
@@ -49,8 +50,8 @@ fn square_jac() {
         polynom::dsquare,
         nalgebra::DVector::from_vec(vec![1.0]),
         nalgebra::DVector::from_vec(vec![2_f64.sqrt()]),
-        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(
+            UpdateQuasiNewtonMethod::GreenstadtFirstMethod,
         )),
         damping,
     );
@@ -66,8 +67,8 @@ fn root_with_high_derivative_jac() {
         polynom::root_with_high_derivative_jac,
         nalgebra::DVector::from_vec(vec![0.15]),
         nalgebra::DVector::from_vec(vec![0.1]),
-        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::InverseJacobianUpdate(
-            UpdateQuasiNewtonMethod::BroydenSecondMethod,
+        ResolutionMethod::QuasiNewton(QuasiNewtonMethod::JacobianUpdate(
+            UpdateQuasiNewtonMethod::GreenstadtFirstMethod,
         )),
         damping,
     );